+++ src/example.test.ts
@@ -13,14 +13,9 @@ function byMethodAndUrl(method: Method, url: string): QueryDefinition<Request> {
   return {
     queryAll: (items) =>
       items.filter((item) => item.method === method && item.url.includes(url)),
-    onNoneFound: (all) =>
-      new Error(
-        `Could not find item with method ${method} and URL containing ${url}. All requests:\n\n${JSON.stringify(all, null, 2)}`,
-      ),
-    onMultipleFound: (all, found) =>
-      new Error(
-        `Multiple items found method ${method} and URL containing ${url}. Found:\n\n${JSON.stringify(found, null, 2)}`,
-      ),
+    serializeForErrorMessage: (item) => JSON.stringify(item, null, 2),
+    noneFoundMessage: `Could not find requests with method ${method} and URL containing ${url}.`,
+    multipleFoundMessage: `Multiple requests found method ${method} and URL containing ${url}.`,
   };
 }